Add 'len' attribute to edges in matrix visualization
The 'len' attribute has been added to the edges in the matrix visualization script, pyarchinit_matrix_exp.py. This change accounts for a more precise representation of the edge length in the generated graphics. The incorrect attribute 'constarint' has also been corrected. These modifications improve the accuracy and quality of matrix visualizations in archaeological data representation.