Add default value to comboBox_sito
The update includes a check that confirms the existence of items in the site_values_list. If the list is not empty, the current text of comboBox_sito is set to the first item of the list. If the list is empty, a default value is set to avoid potential errors or unexpected behaviors.