Update opencv-python version and add directory removal for MacOS
The opencv-python version citation in requirements.txt has been updated to 4.10.0.84. In addition to that, a cleanup process has been added to the __init__.py file, which removes a specific opencv directory if it exists on MacOS systems. This process will help in avoiding potential version conflicts or other related issues.